pyproject.toml
[project]
name = "lectrick"
version = "0.1.0"
description = "Lectrick is an esoteric programming language."
requires-python = ">=3.11,<3.14"
dependencies = [
    "click>=8.2.1",
    "fonttools>=4.58.5",
    "numpy>=2.3.1",
    "pillow>=11.3.0",
    "regex>=2024.11.6",
]

[project.scripts]
main = "lectrick.__main__:main"

[tool.uv]
package = true

[tool.setuptools]
packages = ["lectrick"]

